Rename merge_lidarseg_and_tracking.py to merge_for_panoptic.py
